feat: added public and auth user profile
