+++ src/components/profile/AuthUserProfile.tsx
@@ -1,0 +1,13 @@
+import { UserType } from "../../types";
+import { useAuthContext } from "../context/AuthContext";
+import ProfileLayout from "./ProfileLayout";
+
+const AuthUserProfile = () => {
+    const { token, user } = useAuthContext();
+
+  return (
+    <ProfileLayout user={user} userType={UserType.AUTHENTICATED_USER} />
+  );
+};
+
+export default AuthUserProfile;
