+++ src/components/repository/PublicRepositories.tsx
@@ -13,10 +13,10 @@ const PublicRepositories = () => {
   useEffect(() => {
     const getRepositories = async () => {
       try {
-        // const repositories: Repository[] = await invoke(
-        //   "get_public_repositories"
-        // );
-        const repositories: Repository[] = [];
+        const repositories: Repository[] = await invoke(
+          "get_public_repositories"
+        );
+        // const repositories: Repository[] = [];
         console.log("Public Repositories", repositories);
         setRepositories(repositories);
       } catch (error) {
